Fix UsersAppConfig location
The entry in settings.INSTALLED_APPS could not be changed to "corehq.apps.users.UsersAppConfig" because UsersMiddleware relies on the entry to equal or end with "users". Therefore the app config is being moved where Django expects it to be.
